moving AreaGraph into a component
